remove control-logic hierarchical parent sheet
Sheet is unnecessary for system overview and hides details of the structure, that should be presented.
Removing the sheet also allows to make a clear separation of the power electronics, controller and measurement for
multi PCB Layout. This separation would be hidden away with the control-logic super sheet and would cut through
multiple hierarchical layers.